trading_engine/requirements.txt
alpaca-py
requests
pandas
polars
numpy
scipy
structlog
psycopg2-binary
sqlalchemy
timescaledb
python-dotenv
hmmlearn
pymc
filterpy
statsmodels
vectorbt
PyPortfolioOpt
ollama
apscheduler
pytest
